transformProjectedTypesStream
