+++ src/utils/apply-import-types-and-project.spec.js
@@ -1,0 +1,8 @@
+import apply from './apply-import-types-and-projection';
+
+describe('apply-import-types-and-projection', () => {
+  it('should include all fields by default');
+  it('should remove an unchecked path');
+  it('should deserialize strings to selected types');
+  it('should handle nested objects');
+});
